+++ lib/moon/components/button.ex
@@ -86,7 +86,7 @@ defmodule Moon.Components.Button do
         "w-full bg-hit": @full_width && slot_assigned?(:right_icon_slot),
         "opacity-30": @disabled,
         "anim-pulse animate-[pulse2_1.5s_infinite]": @animation == "pulse",
-        "bg-chi-chi text-goten animate-[error_0.82s_cubic-bezier(0.36,0.07,0.19,0.97)_1_both] anim-error":
+        "bg-chichi text-goten animate-[error_0.82s_cubic-bezier(0.36,0.07,0.19,0.97)_1_both] anim-error":
           @animation == "error"
       }
       disabled={@disabled}